[ot] hw/opentitan: ot_hmac: Make HMAC msg_length / digest writable & implement key swap
This commit adds the checks and functionality for writing to the HMAC
MESSAGE_LENGTH and DIGEST registers whilst the HMAC is not active and
the SHA Engine is disabled, to permit the restoration of saved contexts
(partial computations) using the STOP and CONTINUE commands.

Also implements the KEY_SWAP config option, which now makes the default
configuration use little endian key values, and allows you to use big
endian key values as before by setting the relevant config field to 1.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>