[ot] hw/opentitan: ot_hmac: Remove remaining unnecessary TODO
This removes the last remaining TODO from `ot_hmac` (after the
multi-mode SHA384/SHA512 support is added). Looking at the RTL,
it wipes the secret key regardless of whether the engine is
active or not, even if mid-operation. So this TODO can be
safely removed to mimic the real hardware.

Also fixes a comment line length issue from a previous commit.