[ot] hw/opentitan: ot_rom_ctrl: Remove S&P from data scrambling
In Earlgrey, the S&P layer has been removed from the data scrambling
logic, which affects the scrambling of the generated ROM. The
functionality is removed entirely from the ROM Control to improve error
detection guarantees, interactions with ECC, and timing.

This commit removes the now-redundant logic related to the S&P layer,
instead directly unscrambling words by XORing the keystream and the
`scr`/`in` data.