[ot] hw/opentitan: Use QEMU IRQs instead of Ibex IRQs for alerts
Ibex IRQs are defined using the `ibex_irq_set` function such that they
are optimised to only propagate the IRQ signal if there has been a
change in the level (i.e. they are edge-triggered). Despite this, it is
possible for alerts to be sent from harwdare without previously
resetting the alert, for example upon a hardware fault. This behaviour
can be observed in several OpenTitan tests which force alerts several
times without resetting the state of the block in-between.

One possible solution would just be to send a signal with level=1
followed by a signal with level=0, but the more efficient and
maintainable solution is just to use the original `qemu_irq` for this
purpose, as the IbexIRQ wrapping functionality does not make sense in
this case.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>