[ot] hw/opentitan: ot_aes: fix IV updates in CFB mode
In AES CFB mode the output IV is the ciphertext. For encryption that
corresponds to the block output XORed with the plaintext. See the
block diagrams:

https://en.wikipedia.org/wiki/File:CFB_encryption.svg
https://en.wikipedia.org/wiki/File:CFB_decryption.svg

LibTomCrypt instead provides the output of the block operation (without
the XOR), regardless of whether you use `cfb.IV` or `cfb_getiv`, so we
avoid using the library to update the AES block IV registers.

This fix allows the aes_functest to pass.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>