[ot] hw/riscv: ot_earlgrey: Connect alert escalation severities
Connects GPIO signals for the four different escalation severities of
OpenTitan Earlgrey's Alert handler.

For now, escalation signal 3 (phase 3) connects to the pwrmgr, where it
causes a shutdown as is the default in Darjeeling. In reality, the alert
handler should cause a reset in phase 3 and populate the rstmgr's
`reset_info` accordingly, but this requires additional work in other
blocks, and as such we simply leave this connected to the pwrmgr
shutdown for now to match Darjeeling's functionality.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>