[ot] target/riscv: Add support for Non-Maskable Interrupts
Adds support for Non-Maskable Interrupts (NMIs) to the RISC-V QEMU
CPU implementation. NMIs are defined by the RISC-V specification to be
independent to regular interrupts, used only for hardware error
conditions and causing an immediate jump to an implementation-defined
NMI vector running in M-mode, regardless of the values of e.g. MIE/MIP.

The spec states that "the values written to mcause on an NMI are
implementation-defined"; in Ibex we always use an mcause of 31 (and a
corresponding mtvec offset of 0x7C to determine the jump address for the
NMI handler). To aid a more generic implementation, this commit
introduces a generic `nmi_cause` CPU environment value that can be used
alongside the NMI functionality to allow implementations to change the
mcause as they see fit.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>