[ot] hw/opentitan: ot_alert: Connect alert phase 0 NMI
In escalation phase 0, the alert handler sends an NMI to the CPU, which
will interrupt execution so long as the NMI is enabled in Ibex itself.
As such, we connect this functionality in the alert handler.

See: https://opentitan.org/book/sw/device/silicon_creator/rom/doc/shutdown.html#alert-escalation-phase-actions

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>