[ot] hw/intc: Only clear pending IRQs at PLIC when claimed
See the comment for more details; to allow for proper NMI functionality
based on regular IRQ and NMI handling routines, we must ensure that in
`sifive_plic_irq_request`, we only clear an IRQs pending status when
that IRQ has been claimed. Failure to do so can mean we clear an IRQ at
a peripheral as part of an NMI service routine and as such never call
into the IRQ handler when the NMI handler returns, as would be expected.
This commit updates the SiFive PLIC so that it works as expected.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>