[ot] hw/opentitan: ot_flash: Fix flash disable behaviour
In the existing flash, if you write to the `DIS` register to disable the
flash, it completely removes the ability to read from or write to any of
the CSRs.

Referencing the documentation:
https://opentitan.org/book/hw/top_earlgrey/ip_autogen/flash_ctrl/doc/theory_of_operation.html#flash-access-disable

Upon disabling the flash, it should instead finish existing stateful
operations, and then give a memory protection error for any further
flash protocol controller initiated operations. That is; you should
still be able to interact with the registers, but no new flash
transactions can be started via the Flash Controller. The host-facing
TL-UL adapter errors back all host initiated operations when disabled,
so the disabling the relevant mem region is kept.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>