[ot] hw/opentitan: clkmgr,spi_host: earlgrey-1.0.0 backcompat defs
Earlgrey 1.0.0 uses SPI Host V2, not V3, which has breaking changes to
the size and locations of the field in the SPI Host command register.
This can just be adapted by changing the field definitions; the command
length in earlgrey-1.0.0 is smaller and so the rest of the code is
backwards-compatible.

For the clkmgr, there was a known RTL bug in 1.0.0 where the jitter
enable register was not actually protected by its respective REGWEN
register. This replicates that behaviour for correct testing on the
earlgrey-1.0.0 branch.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>