[ot] docs/opentitan: ot_earlgrey: Update documentation
Change the supported version to be `earlgrey_1.0.0` instead of
`2.5.2-RC0` - the former is a stable silicon release whereas the latter
contains a variety of outdated hardware and bugs which does not make as
much sense as an emulation target.

This changes the supported version of the `ot_earlgrey` machine to be
Earlgrey 1.0.0 specifically.

Also add a small amount more documentation for some updated IP that was
missed, and more completely enumerate the current list of IP that are
not supported in QEMU Earlgrey emulation.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>