[ot] hw/opentitan: ot_keymgr: Initial keymgr FSM with reset state
This commit introduces the initial logic comrpising the keymgr FSM,
stubbing out all the different internal FSM states apart from the reset
state, which is implemented.

Note however that the reset state will not currently function because
kwymgr enablement via the lc_ctrl connection has not yet been added.

Add logic for scheduling the FSM with a BottomHalf to appropriately
defer execution to the end of I/O calls. Stub out all the different FSM
states but add functionality for changing the internal working state of
the keymgr. Also add logic on writes to the `START` register to tick the
main FSM.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>