[ot] scripts/opentitan: cfggen.py: Extract RndCnstCdi from Keymgr HJSON
In the OpenTitan keymgr HJSON, there is a parameter "RndCnstCdi"
containing the compile-time random bits for the generation seed when no
CDI is selected. This is like "RndCnstNoneSeed" (also present in
keymgr_dpe), but used when there is no CDI selected instead of no
destination selected.

Unlike the other compile-time random seeds, this is not suffixed with
"Seed" and thus is not extracted by the existing regex.

This commit adds an additional regex to extract this attribute.
Ideally it would be given a more appropriate name (e.g. "cdi_seed" or
"no_cdi_gen_seed") along with a few other of these parameter names, but
that can be deferred to a later change.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>