[ot] hw/riscv: ot_earlgrey: Connect & stub LC signals
Due to limited support for other HW peripherals, Earlgrey is missing
all of its life cycle broadcast signals. Most notably, the outdated OTP
controller uses a lot of these broadcast signals.

Where possible, update the Earlgrey implementation to add the LC
broadcast signals to achieve parity with the Darjeeling implementation.
To do this, implement IRQ splitters for certain LC signals, connect up
the LC_CTRL with the ibex_wrapper and pwrmgr, and most importantly,
leave TODOs noting the signals that are not connected and why.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>