[ot] hw/riscv: ot_earlgrey: Add signals between lc_ctrl and otp_ctrl
Now that Earlgrey's OTP is updated to match Darjeeling and can support
broadcast signals from the lifecycle controller, hook up the two
broadcast signals that are relevant on Earlgrey (CREATOR_SEED and
OWNER_SEED are not stored in OTP on EG, unlike DJ, and so there is no
need for connections with the OTP here). Also connects the escalation
signal, though this has limited suppport in the otp_ctrl at present.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>