[ot] hw/riscv: ot_earlgrey: Add signals between lc_ctrl and otp_ctrl
Now that Earlgrey's OTP is updated to match Darjeeling and can support
broadcast signals from the lifecycle controller, hook up the three
broadcast signals that are relevant on Earlgrey (OWNER_SEED is not
stored in OTP on EG, unlike DJ, and so there is no need for connections
with the OTP here). Also connects the escalation signal, though this
has limited support in the otp_ctrl at present.

The `OT_LC_CREATOR_SEED_SW_RW_EN` also needs to be connected to the OTP.
Even though it sounds like this should only sound the Software read /
writability of the Creator Seed (stored in flash on Earlgrey), it
actually controls any keymgr_creator partition, which includes the
SECRET2 partition containing the Creator Root Key. So this is maybe more
aptly a "CREATOR_SECRET_SW_RW_EN" seed, as it pertains to many creator
secrets and not just the Creator Seed.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>