[ot] hw/opentitan: ot_keymgr: Add root key loading from OTP
Add a connection between the keymgr and the otp_ctrl, and use the
implement otp_ctrl methods to retrieve the creator root key from OTP
upon a KeyMgr advance operation.

Change the behavior of the ROOT_KEY FSM state so that an invalid root
key or disablement will cause a transition to the WIPE state, and
implement internal key state storage to store the loaded keys.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>