[ot] hw/opentitan: ot_otp_dj: Fix generated SRAM key writeback
Currently the generated SRAM scrambling key is not written back to the
key->seed field correctly: for each scrambling block (2 for the SRAM
key), we are instead just taking the LSB 8 bytes of the generated 64bit
seed and writing it in that index. This means that the majority of the
key will still contain the initialized values, and 56 bits of each
generated key is ignored.

Correct the writeback logic, iterating over each byte in the scrambling
block and individually writing each byte.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>