[ot] hw/opentitan: ot_keymgr_dpe: Clear cfg_regwen when disabled by LC
To more closely match the RTL, the CFG_REGWEN value should also be
disabled when the keymgr itself is disabled by the lc_ctrl (as well as
during a keymgr operation).

Also, take the opportunity to refactor a large compound expression
slightly to make the code a bit easier to read.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>