[ot] hw/opentitan: ot_keymgr: Add operation status & init FSM state
This commit is mainly focused on creating a (mostly complete)
implementation of the `Init` keymgr FSM state which performs some checks
before starting an advance/disable operation. These operations are both
stubbed out for now, to be implemented in the future.

As part of this change, logic (which is the same as the keymgr_dpe
logic) is added for updating the `op_status` to latch for and reflect
the last requested operation. Additionally, functionality is added to
lock the CFG_REGWEN during an operation.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>