[ot] hw/opentitan: ot_keymgr: Add operation status & init FSM state
This commit focuses on creating a (mostly complete) implementation of
the `INIT` keymgr FSM state, which performs some checks before starting
an advance or disable operation. These are both stubbed out for now.

Logic (which is the same as the keymgr_dpe logic) is added for updating
the `OP_STATUS`, so that it latches and reflects the last requested
operation. Also, functionality is added to lock `CFG_REGWEN` during an
operation.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>