[ot] hw/opentitan: ot_flash: Add keymgr seed retrieval
Fill out more of the flash controller initialization logic, now
implementing the retrieval of the keymgr (owner and creator) seeds from
flash during initialization, which is the first instance of a hardware
request being used through the flash controller model. These keys are
loaded into the flash controller and stored until they are requested by
the keymgr, though this is not hooked up to the keymgr currently.

Also performs a few more sanity checks with some errors: check that we
aren't already initialized (or initializing) when we try to initialize.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>