[ot] hw/opentitan: ot_flash: Add lc broadcast page permissions & escalation logic
This commit introduces the logic for processing many of the LC broadcast
signals in the flash controller. The broadcast logic is handled using
the same method as the existing OTP implementation, storing the
broadcast signals and deferring their handling with a BH to decouple the
signaling from the handling.

The `LC_SEED_HW_RD_EN` should gate reading the seeds during
initialisation, but this is not implemented yet as I am not sure the
lc_ctrl is developed enough to support this yet. The `LC_NVM_DEBUG_EN`
broadcast signal for JTAG connections also is unimplemented.

The four signals controlling the qualified info pages storing flash
secrets (and the isolated partition) are all connected, and the
escalation signal is partially implemented, disabling the flash for now.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>