[ot] hw/opentitan: ot_flash: Add device ID to traces and logs
To bring the flash controller more in line with other modernized QEMU
IP, add an `ot_id` property which is added as context to all traces and
logs. This is primarily useful if you have multiple instances of a block
(e.g. if we had a future top with multiple flash controllers) or if you
want to rename a device, but is added regardless to be consistent with
other OpenTitan QEMU devices.

All traces are updated to start with the ID, and all of the qemu masked
logs and errors also include the ID.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>