[ot] hw/opentitan: ot_keymgr: Add KMAC KDF advance input computation
This commit introduces the logic relevant to computing the KDF `AdvData`
input for the `Advance` operation, which depends on both the current
keymgr stage and the CDI.

This logic is adapted from the keymgr_dpe, but with many differences:
 - `keymgr` has no slots, and encodes stages in its main FSM states.
 - `keymgr` has 2 CDIs: one for sealing and one for attestation.
 Normally, you can select the CDI, but advance operations advance
 both (sealing, then attestation).
 - `keymgr` uses different KDF inputs to `keymgr_dpe`. It only has 1
 ROM digest, uses creator/owner seeds in different phases, and
 concatenates its inputs in a different order.
 - `keymgr` gets creator/owner seeds from flash, instead of OTP.
 - `keymgr` chooses to use either sealing or attestation SW bindings
   based on the CDI, `keymgr_dpe` only has 1 SW binding.

For now, many parts are still stubbed out - we do not yet send the KDF
request to KMAC, and we do not properly handle wiping data on
disablement, or invalid inputs.

This commit also adds minimal handling for the `disable` operation.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>