[ot] HACK: Allow executing pyot.py without any tests loaded from a pyot test config file.
Currently the code assumes that test executions must be from the
pyot.hjson file. You can try and manually give an app, but this part of
the code seems to have rotted due to lack of use and the app definition
built is woefully lacking compared to the execution context, which
requires something that at least "looks like a test".

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>