[ot] hw/opentitan: ot_keymgr: Restore KMAC sideload key after KDF
Correct the behaviour of the keymgr to restore the previous output KMAC
sideloading key after finishing any keymgr operation that uses KMAC (for
offloading KDF). The sideloading HW output is not yet implemented, but
this functionality will be used in future commits which add this.

As per the Keymgr docs: "... The KMAC key thus has two possible outputs,
one is the sideload key, and the other is internal state key. When a
valid operation is called, the internal state key is sent over the KMAC
key. During all other times, the sideloaded value is presented..."

Implement this in QEMU by just saving the current sideload key to some
internal "backup" state on every sideload key push, and after completing
an operation we restore this saved key (instead of clearing the sink).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>