[ot] hw/opentitan: ot_otp: validate presence of all scrambling keys at init time.
Scrambling keys are now a mandatory feature if an OTP file is loaded.
Early abort and warn the user if a key is found missing (from the submitted
configuration file)

Report the partition name (which should match the secret key)

When no OTP file is provided, i.e. no OTP data is loaded, unscrambling stage is
not executed; scrambling keys are neither used nor required.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>