[ot] hw/opentitan: ot_otp_eg: Fix buffered DAI reads for scrambled parts
DAI reads from buffered partitions should actually read from the buffer
and not just always the OTP data, to reflect what real hardware is
actually doing. This was never a problem before because, without
scrambling, the OTP data should always match what is in the buffer (and
software never pays the price of "slowly" reading OTP storage).

However, now that scrambling is introduced, the descrambled data is
stored in the buffer. Thus, when a DAI read fetches from OTP, it is
actually fetching the scrambled data - this matches the ECC
(calculated over the scrambled data), but does not return the correct
unscrambled value.

To correct this, change the OTP so that DAI reads of buffered partitions
actually read from the buffer and not from the OTP storage. Similarly,
disable ECC checks on reads of buffered partitions; the ECC was
calculated and checked during the original OTP read, it does not need to
be recalculated.

(aside: there does seem to be an additional ECC protection for buffered
partitions as part of the integrity check which will check the ECC and
computed digest, but this is separate to the ECC on OTP reads).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>