[ot] target/riscv: ibex_csr: Fix Ibex `mtvec` WARL writes
This fixes the behaviour of mtvec writes on Ibex to allow writes in
non-vectored mtvec encode modes, but always tie the written mode to
vectored (mtvec is WARL).

This is required for the ECALL RISC-V compliance test to pass in QEMU
(as it currently does on FPGA and Silicon targets).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>